refactor linter
