+++ linter/oprator_test.go
@@ -1,0 +1,412 @@
+package linter
+
+import "testing"
+
+func TestLintEqualOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host == "example.com") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host == "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host == 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+
+	t.Run("req.backend is comparable with BACKEND type", func(t *testing.T) {
+		input := `
+backend foo {}
+sub foo {
+	if (req.backend == foo) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+}
+
+func TestLintNotEqualOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host != "example.com") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host != "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host != 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+}
+
+func TestLintGreaterThanOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I > 10) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host > "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host > 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare INTEGER vs FLOAT", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I > 10.0) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+}
+
+func TestLintGreaterThanEqualOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I >= 10) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host >= "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host >= 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare INTEGER vs FLOAT", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I >= 10.0) {
+		restart;
+	}
+}`
+		assertError(t, input)
+
+	})
+}
+
+func TestLintLessThanOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I < 10) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host < "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host < 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare INTEGER vs FLOAT", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I < 10.0) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+}
+
+func TestLintLessThanEqualOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I <= 10) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host <= "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host <= 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare INTEGER vs FLOAT", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.I INTEGER;
+	set var.I = 100;
+	if (var.I <= 10.0) {
+		restart;
+	}
+}`
+		assertError(t, input)
+
+	})
+}
+
+func TestLintRegexOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host ~ "example") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("pass with acl", func(t *testing.T) {
+		input := `
+acl internal {
+	"10.0.0.10";
+}
+
+sub foo {
+	if (req.http.Host ~ internal) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host ~ "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host ~ 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+
+	t.Run("pass with PCRE expression", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host ~ "(?i)^word") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("pass with expression that has backslash", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.User-Agent ~ "\(compatible.?; Googlebot/2.1.?; \+http://www.google.com/bot.html") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("pass with PCRE expression that has backslash", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.User-Agent ~ "(?i)windows\ ?ce") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("pass with PCRE expression that uses atomic grouping (unsupported by regexp)", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.User-Agent ~ "\b(?>integer|insert|in)\b") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+}
+
+func TestLintRegexNotOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host !~ "example") {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("pass with acl", func(t *testing.T) {
+		input := `
+acl internal {
+	"10.0.0.10";
+}
+
+sub foo {
+	if (req.http.Host !~ internal) {
+		restart;
+	}
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("cannot use in other statement", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.BoolItem BOOL;
+	set var.BoolItem = req.http.Host !~ "example.com";
+}`
+		assertError(t, input)
+	})
+
+	t.Run("cannot compare for different type", func(t *testing.T) {
+		input := `
+sub foo {
+	if (req.http.Host !~ 10) {
+		restart;
+	}
+}`
+		assertError(t, input)
+	})
+}
+
+func TestLintPlusOperator(t *testing.T) {
+	t.Run("pass", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.S STRING;
+	set var.S = "foo" "bar" + "baz";
+}`
+		assertNoError(t, input)
+	})
+
+	t.Run("raise warning concatenation without string type", func(t *testing.T) {
+		input := `
+sub foo {
+	declare local var.S STRING;
+	declare local var.I INTEGER;
+
+	set var.I = 10;
+	set var.S = "foo" "bar" + var.I;
+}`
+		// error, but warning
+		assertErrorWithSeverity(t, input, INFO)
+	})
+}
